+++ sdk/consumption/arm-consumption/src/models/index.ts
@@ -12,6 +12,10 @@ export type LegacyReservationRecommendationPropertiesUnion =
   | LegacyReservationRecommendationProperties
   | LegacySingleScopeReservationRecommendationProperties
   | LegacySharedScopeReservationRecommendationProperties;
+export type ModernReservationRecommendationPropertiesUnion =
+  | ModernReservationRecommendationProperties
+  | ModernSingleScopeReservationRecommendationProperties
+  | ModernSharedScopeReservationRecommendationProperties;
 export type UsageDetailUnion =
   | UsageDetail
   | LegacyUsageDetail
@@ -136,8 +140,6 @@ export interface BudgetTimePeriod {
 export interface BudgetFilter {
   /** The logical "AND" expression. Must have at least 2 items. */
   and?: BudgetFilterProperties[];
-  /** The logical "NOT" expression. */
-  not?: BudgetFilterProperties;
   /** Has comparison expression for a dimension */
   dimensions?: BudgetComparisonExpression;
   /** Has comparison expression for a tag */
@@ -343,7 +345,7 @@ export interface ReservationRecommendationDetailsResourceProperties {
    */
   readonly appliedScopes?: string[];
   /**
-   * On demand rate of the resource.
+   * Hourly on-demand rate of the resource. Includes only hardware rate i.e, software rate is not included.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly onDemandRate?: number;
@@ -358,7 +360,7 @@ export interface ReservationRecommendationDetailsResourceProperties {
    */
   readonly region?: string;
   /**
-   * Reservation rate of the resource.
+   * Hourly reservation rate of the resource. Varies based on the term.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly reservationRate?: number;
@@ -400,15 +402,15 @@ export interface ReservationRecommendationDetailsSavingsProperties {
   readonly unitOfMeasure?: string;
 }
 
-/** Details of estimated savings. */
+/** Details of estimated savings. The costs and savings are estimated for the term. */
 export interface ReservationRecommendationDetailsCalculatedSavingsProperties {
   /**
-   * The cost without reservation.
+   * The cost without reservation. Includes hardware and software cost.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly onDemandCost?: number;
   /**
-   * The difference between total reservation cost and reservation cost.
+   * Hardware and software cost of the resources not covered by the reservation.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly overageCost?: number;
@@ -418,19 +420,19 @@ export interface ReservationRecommendationDetailsCalculatedSavingsProperties {
    */
   readonly quantity?: number;
   /**
-   * The exact cost of the estimated usage using reservation.
+   * Hardware cost of the resources covered by the reservation.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly reservationCost?: number;
   /**
-   * The cost of the suggested quantity.
+   * Reservation cost + software cost of the resources covered by the reservation + overage cost.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly totalReservationCost?: number;
   /** The number of reserved units used to calculate savings. Always 1 for virtual machines. */
   reservedUnitCount?: number;
   /**
-   * The amount saved by purchasing the recommended quantity of reservation.
+   * The amount saved by purchasing the recommended quantity of reservation. This is equal to onDemandCost - totalReservationCost.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly savings?: number;
@@ -889,6 +891,92 @@ export interface SkuProperty {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly value?: string;
+}
+
+/** The properties of the reservation recommendation. */
+export interface ModernReservationRecommendationProperties {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  scope: "Single" | "Shared";
+  /**
+   * Resource Location.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly location?: string;
+  /**
+   * The number of days of usage to look back for recommendation.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly lookBackPeriod?: number;
+  /**
+   * The instance Flexibility Ratio.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly instanceFlexibilityRatio?: number;
+  /**
+   * The instance Flexibility Group.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly instanceFlexibilityGroup?: string;
+  /**
+   * The normalized Size.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly normalizedSize?: string;
+  /**
+   * The recommended Quantity Normalized.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly recommendedQuantityNormalized?: number;
+  /**
+   * The meter id (GUID)
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly meterId?: string;
+  /**
+   * RI recommendations in one or three year terms.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly term?: string;
+  /**
+   * The total amount of cost without reserved instances.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly costWithNoReservedInstances?: Amount;
+  /**
+   * Recommended quality for reserved instances.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly recommendedQuantity?: number;
+  /**
+   * Resource type.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly resourceType?: string;
+  /**
+   * The total amount of cost with reserved instances.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly totalCostWithReservedInstances?: Amount;
+  /**
+   * Total estimated savings with reserved instances.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly netSavings?: Amount;
+  /**
+   * The usage date for looking back.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly firstUsageDate?: Date;
+  /**
+   * List of sku properties
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly skuProperties?: SkuProperty[];
+  /**
+   * This is the ARM Sku name.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly skuName?: string;
 }
 
 /** The properties of the price sheet download. */
@@ -1369,50 +1457,6 @@ export interface ManagementGroupAggregatedCostResult extends Resource {
   excludedSubscriptions?: string[];
 }
 
-/** A credit summary resource. */
-export interface CreditSummary extends Resource {
-  /**
-   * Summary of balances associated with this credit summary.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly balanceSummary?: CreditBalanceSummary;
-  /**
-   * Pending credit adjustments.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly pendingCreditAdjustments?: Amount;
-  /**
-   * Expired credit.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly expiredCredit?: Amount;
-  /**
-   * Pending eligible charges.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly pendingEligibleCharges?: Amount;
-  /**
-   * The credit currency.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly creditCurrency?: string;
-  /**
-   * The billing currency.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly billingCurrency?: string;
-  /**
-   * Credit's reseller.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly reseller?: Reseller;
-  /**
-   * The eTag for the resource.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly eTag?: string;
-}
-
 /** A budget resource. */
 export interface Budget extends ProxyResource {
   /** The category of the budget, whether the budget tracks cost or usage. */
@@ -1651,6 +1695,50 @@ export interface LotSummary extends ProxyResource {
   readonly eTagPropertiesETag?: string;
 }
 
+/** A credit summary resource. */
+export interface CreditSummary extends ProxyResource {
+  /**
+   * Summary of balances associated with this credit summary.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly balanceSummary?: CreditBalanceSummary;
+  /**
+   * Pending credit adjustments.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly pendingCreditAdjustments?: Amount;
+  /**
+   * Expired credit.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly expiredCredit?: Amount;
+  /**
+   * Pending eligible charges.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly pendingEligibleCharges?: Amount;
+  /**
+   * The credit currency.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly creditCurrency?: string;
+  /**
+   * The billing currency.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly billingCurrency?: string;
+  /**
+   * Credit's reseller.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly reseller?: Reseller;
+  /**
+   * The eTag for the resource.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly eTagPropertiesETag?: string;
+}
+
 /** Reservation transaction resource. */
 export interface ReservationTransaction extends ReservationTransactionResource {
   /**
@@ -1669,7 +1757,7 @@ export interface ReservationTransaction extends ReservationTransactionResource {
    */
   readonly description?: string;
   /**
-   * The type of the transaction (Purchase, Cancel, etc.)
+   * The type of the transaction (Purchase, Cancel or Refund).
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly eventType?: string;
@@ -1814,7 +1902,7 @@ export interface ModernReservationTransaction
    */
   readonly eventDate?: Date;
   /**
-   * The type of the transaction (Purchase, Cancel, etc.)
+   * The type of the transaction (Purchase, Cancel or Refund).
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly eventType?: string;
@@ -1904,6 +1992,25 @@ export interface LegacySingleScopeReservationRecommendationProperties
 /** The properties of the legacy reservation recommendation for shared scope. */
 export interface LegacySharedScopeReservationRecommendationProperties
   extends LegacyReservationRecommendationProperties {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  scope: "Shared";
+}
+
+/** The properties of the modern reservation recommendation for single scope. */
+export interface ModernSingleScopeReservationRecommendationProperties
+  extends ModernReservationRecommendationProperties {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  scope: "Single";
+  /**
+   * Subscription ID associated with single scoped recommendation.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly subscriptionId?: string;
+}
+
+/** The properties of the modern reservation recommendation for shared scope. */
+export interface ModernSharedScopeReservationRecommendationProperties
+  extends ModernReservationRecommendationProperties {
   /** Polymorphic discriminator, which specifies the different types this object can be */
   scope: "Shared";
 }
@@ -2127,6 +2234,16 @@ export interface LegacyUsageDetail extends UsageDetail {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly payGPrice?: number;
+  /**
+   * Unique identifier for the applicable benefit.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly benefitId?: string;
+  /**
+   * Name of the applicable benefit.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly benefitName?: string;
   /**
    * Identifier that indicates how the meter is priced.
    * NOTE: This property will not be serialized. It can only be populated by the server.
@@ -2635,6 +2752,11 @@ export interface ModernReservationRecommendation
    */
   readonly recommendedQuantity?: number;
   /**
+   * Resource type.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly resourceType?: string;
+  /**
    * The total amount of cost with reserved instances.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
@@ -2649,11 +2771,8 @@ export interface ModernReservationRecommendation
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly firstUsageDate?: Date;
-  /**
-   * Shared or single recommendation.
-   * NOTE: This property will not be serialized. It can only be populated by the server.
-   */
-  readonly scope?: string;
+  /** Shared or single recommendation. */
+  scope: string;
   /**
    * List of sku properties
    * NOTE: This property will not be serialized. It can only be populated by the server.
@@ -2699,7 +2818,7 @@ export interface LegacyChargeSummary extends ChargeSummary {
    * Marketplace Charges.
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
-  readonly marketplaceCharges?: number;
+  readonly azureMarketplaceCharges?: number;
   /**
    * Currency Code
    * NOTE: This property will not be serialized. It can only be populated by the server.
@@ -2766,6 +2885,11 @@ export interface ModernChargeSummary extends ChargeSummary {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly isInvoiced?: boolean;
+  /**
+   * Subscription guid.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly subscriptionId?: string;
 }
 
 /** Legacy Reservation transaction resource. */
@@ -2872,11 +2996,11 @@ export type BudgetOperatorType = string;
 
 /** Known values of {@link OperatorType} that the service accepts. */
 export enum KnownOperatorType {
-  /** EqualTo */
+  /** Alert will be triggered if the evaluated cost is the same as threshold value. Note: It’s not recommended to use this OperatorType as there’s low chance of cost being exactly the same as threshold value, leading to missing of your alert. This OperatorType will be deprecated in future. */
   EqualTo = "EqualTo",
-  /** GreaterThan */
+  /** Alert will be triggered if the evaluated cost is greater than the threshold value. Note: This is the recommended OperatorType while configuring Budget Alert. */
   GreaterThan = "GreaterThan",
-  /** GreaterThanOrEqualTo */
+  /** Alert will be triggered if the evaluated cost is greater than or equal to the threshold value. */
   GreaterThanOrEqualTo = "GreaterThanOrEqualTo"
 }
 
@@ -2885,17 +3009,17 @@ export enum KnownOperatorType {
  * {@link KnownOperatorType} can be used interchangeably with OperatorType,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
- * **EqualTo** \
- * **GreaterThan** \
- * **GreaterThanOrEqualTo**
+ * **EqualTo**: Alert will be triggered if the evaluated cost is the same as threshold value. Note: It’s not recommended to use this OperatorType as there’s low chance of cost being exactly the same as threshold value, leading to missing of your alert. This OperatorType will be deprecated in future.  \
+ * **GreaterThan**: Alert will be triggered if the evaluated cost is greater than the threshold value. Note: This is the recommended OperatorType while configuring Budget Alert. \
+ * **GreaterThanOrEqualTo**: Alert will be triggered if the evaluated cost is greater than or equal to the threshold value.
  */
 export type OperatorType = string;
 
 /** Known values of {@link ThresholdType} that the service accepts. */
 export enum KnownThresholdType {
-  /** Actual */
+  /** Actual costs budget alerts notify when the actual accrued cost exceeds the allocated budget . */
   Actual = "Actual",
-  /** Forecasted */
+  /** Forecasted costs budget alerts provide advanced notification that your spending trends are likely to exceed your allocated budget, as it relies on forecasted cost predictions. */
   Forecasted = "Forecasted"
 }
 
@@ -2904,8 +3028,8 @@ export enum KnownThresholdType {
  * {@link KnownThresholdType} can be used interchangeably with ThresholdType,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
- * **Actual** \
- * **Forecasted**
+ * **Actual**: Actual costs budget alerts notify when the actual accrued cost exceeds the allocated budget . \
+ * **Forecasted**: Forecasted costs budget alerts provide advanced notification that your spending trends are likely to exceed your allocated budget, as it relies on forecasted cost predictions.
  */
 export type ThresholdType = string;
 
@@ -3059,6 +3183,24 @@ export enum KnownReservationRecommendationKind {
  */
 export type ReservationRecommendationKind = string;
 
+/** Known values of {@link Scope} that the service accepts. */
+export enum KnownScope {
+  /** Single */
+  Single = "Single",
+  /** Shared */
+  Shared = "Shared"
+}
+
+/**
+ * Defines values for Scope. \
+ * {@link KnownScope} can be used interchangeably with Scope,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Single** \
+ * **Shared**
+ */
+export type Scope = string;
+
 /** Known values of {@link Term} that the service accepts. */
 export enum KnownTerm {
   /** 1 year reservation term */
@@ -3113,7 +3255,9 @@ export enum KnownEventType {
   /** UnKnown */
   UnKnown = "UnKnown",
   /** NewCredit */
-  NewCredit = "NewCredit"
+  NewCredit = "NewCredit",
+  /** CreditExpired */
+  CreditExpired = "CreditExpired"
 }
 
 /**
@@ -3127,7 +3271,8 @@ export enum KnownEventType {
  * **PendingNewCredit** \
  * **PendingExpiredCredit** \
  * **UnKnown** \
- * **NewCredit**
+ * **NewCredit** \
+ * **CreditExpired**
  */
 export type EventType = string;
 
@@ -3202,24 +3347,6 @@ export enum KnownPricingModelType {
  * **Spot**
  */
 export type PricingModelType = string;
-
-/** Known values of {@link Scope} that the service accepts. */
-export enum KnownScope {
-  /** Single */
-  Single = "Single",
-  /** Shared */
-  Shared = "Shared"
-}
-
-/**
- * Defines values for Scope. \
- * {@link KnownScope} can be used interchangeably with Scope,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Single** \
- * **Shared**
- */
-export type Scope = string;
 
 /** Optional parameters. */
 export interface UsageDetailsListOptionalParams
@@ -3472,7 +3599,7 @@ export type ReservationRecommendationDetailsGetResponse = ReservationRecommendat
 /** Optional parameters. */
 export interface ReservationTransactionsListOptionalParams
   extends coreClient.OperationOptions {
-  /** Filter reservation transactions by date range. The properties/EventDate for start date and end date. The filter supports 'le' and  'ge' */
+  /** Filter reservation transactions by date range. The properties/EventDate for start date and end date. The filter supports 'le' and  'ge'. Note: API returns data for the entire start date's and end date's billing month. For example, filter properties/eventDate+ge+2020-01-01+AND+properties/eventDate+le+2020-12-29 will include data for the entire December 2020 month (i.e. will contain records for dates December 30 and 31) */
   filter?: string;
 }
 
@@ -3482,7 +3609,7 @@ export type ReservationTransactionsListResponse = ReservationTransactionsListRes
 /** Optional parameters. */
 export interface ReservationTransactionsListByBillingProfileOptionalParams
   extends coreClient.OperationOptions {
-  /** Filter reservation transactions by date range. The properties/EventDate for start date and end date. The filter supports 'le' and  'ge' */
+  /** Filter reservation transactions by date range. The properties/EventDate for start date and end date. The filter supports 'le' and  'ge'. Note: API returns data for the entire start date's and end date's billing month. For example, filter properties/eventDate+ge+2020-01-01+AND+properties/eventDate+le+2020-12-29 will include data for entire December 2020 month (i.e. will contain records for dates December 30 and 31) */
   filter?: string;
 }
 
@@ -3611,6 +3738,16 @@ export interface LotsListByBillingAccountOptionalParams
 export type LotsListByBillingAccountResponse = Lots;
 
 /** Optional parameters. */
+export interface LotsListByCustomerOptionalParams
+  extends coreClient.OperationOptions {
+  /** May be used to filter the lots by Status, Source etc. The filter supports 'eq', 'lt', 'gt', 'le', 'ge', and 'and'. Tag filter is a key value pair string where key and value is separated by a colon (:). */
+  filter?: string;
+}
+
+/** Contains response data for the listByCustomer operation. */
+export type LotsListByCustomerResponse = Lots;
+
+/** Optional parameters. */
 export interface LotsListByBillingProfileNextOptionalParams
   extends coreClient.OperationOptions {}
 
@@ -3623,6 +3760,13 @@ export interface LotsListByBillingAccountNextOptionalParams
 
 /** Contains response data for the listByBillingAccountNext operation. */
 export type LotsListByBillingAccountNextResponse = Lots;
+
+/** Optional parameters. */
+export interface LotsListByCustomerNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listByCustomerNext operation. */
+export type LotsListByCustomerNextResponse = Lots;
 
 /** Optional parameters. */
 export interface CreditsGetOptionalParams extends coreClient.OperationOptions {}