+++ sdk/consumption/arm-consumption/src/operations/reservationsSummaries.ts
@@ -220,20 +220,20 @@ export class ReservationsSummariesImpl implements ReservationsSummaries {
 
   /**
    * Lists the reservations summaries for the defined scope daily or monthly grain.
-   * @param scope The scope associated with reservations summaries operations. This includes
-   *              '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for BillingAccount scope (legacy),
-   *              and
-   *              '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}/billingProfiles/{billingProfileId}'
-   *              for BillingProfile scope (modern).
+   * @param resourceScope The scope associated with reservations summaries operations. This includes
+   *                      '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for BillingAccount scope (legacy),
+   *                      and
+   *                      '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}/billingProfiles/{billingProfileId}'
+   *                      for BillingProfile scope (modern).
    * @param grain Can be daily or monthly
    * @param options The options parameters.
    */
   public list(
-    scope: string,
+    resourceScope: string,
     grain: Datagrain,
     options?: ReservationsSummariesListOptionalParams
   ): PagedAsyncIterableIterator<ReservationSummary> {
-    const iter = this.listPagingAll(scope, grain, options);
+    const iter = this.listPagingAll(resourceScope, grain, options);
     return {
       next() {
         return iter.next();
@@ -245,13 +245,13 @@ export class ReservationsSummariesImpl implements ReservationsSummaries {
         if (settings?.maxPageSize) {
           throw new Error("maxPageSize is not supported by this operation.");
         }
-        return this.listPagingPage(scope, grain, options, settings);
+        return this.listPagingPage(resourceScope, grain, options, settings);
       }
     };
   }
 
   private async *listPagingPage(
-    scope: string,
+    resourceScope: string,
     grain: Datagrain,
     options?: ReservationsSummariesListOptionalParams,
     settings?: PageSettings
@@ -259,14 +259,14 @@ export class ReservationsSummariesImpl implements ReservationsSummaries {
     let result: ReservationsSummariesListResponse;
     let continuationToken = settings?.continuationToken;
     if (!continuationToken) {
-      result = await this._list(scope, grain, options);
+      result = await this._list(resourceScope, grain, options);
       let page = result.value || [];
       continuationToken = result.nextLink;
       setContinuationToken(page, continuationToken);
       yield page;
     }
     while (continuationToken) {
-      result = await this._listNext(scope, continuationToken, options);
+      result = await this._listNext(resourceScope, continuationToken, options);
       continuationToken = result.nextLink;
       let page = result.value || [];
       setContinuationToken(page, continuationToken);
@@ -275,11 +275,15 @@ export class ReservationsSummariesImpl implements ReservationsSummaries {
   }
 
   private async *listPagingAll(
-    scope: string,
+    resourceScope: string,
     grain: Datagrain,
     options?: ReservationsSummariesListOptionalParams
   ): AsyncIterableIterator<ReservationSummary> {
-    for await (const page of this.listPagingPage(scope, grain, options)) {
+    for await (const page of this.listPagingPage(
+      resourceScope,
+      grain,
+      options
+    )) {
       yield* page;
     }
   }
@@ -324,21 +328,21 @@ export class ReservationsSummariesImpl implements ReservationsSummaries {
 
   /**
    * Lists the reservations summaries for the defined scope daily or monthly grain.
-   * @param scope The scope associated with reservations summaries operations. This includes
-   *              '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for BillingAccount scope (legacy),
-   *              and
-   *              '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}/billingProfiles/{billingProfileId}'
-   *              for BillingProfile scope (modern).
+   * @param resourceScope The scope associated with reservations summaries operations. This includes
+   *                      '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for BillingAccount scope (legacy),
+   *                      and
+   *                      '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}/billingProfiles/{billingProfileId}'
+   *                      for BillingProfile scope (modern).
    * @param grain Can be daily or monthly
    * @param options The options parameters.
    */
   private _list(
-    scope: string,
+    resourceScope: string,
     grain: Datagrain,
     options?: ReservationsSummariesListOptionalParams
   ): Promise<ReservationsSummariesListResponse> {
     return this.client.sendOperationRequest(
-      { scope, grain, options },
+      { resourceScope, grain, options },
       listOperationSpec
     );
   }
@@ -384,21 +388,21 @@ export class ReservationsSummariesImpl implements ReservationsSummaries {
 
   /**
    * ListNext
-   * @param scope The scope associated with reservations summaries operations. This includes
-   *              '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for BillingAccount scope (legacy),
-   *              and
-   *              '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}/billingProfiles/{billingProfileId}'
-   *              for BillingProfile scope (modern).
+   * @param resourceScope The scope associated with reservations summaries operations. This includes
+   *                      '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for BillingAccount scope (legacy),
+   *                      and
+   *                      '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}/billingProfiles/{billingProfileId}'
+   *                      for BillingProfile scope (modern).
    * @param nextLink The nextLink from the previous successful call to the List method.
    * @param options The options parameters.
    */
   private _listNext(
-    scope: string,
+    resourceScope: string,
     nextLink: string,
     options?: ReservationsSummariesListNextOptionalParams
   ): Promise<ReservationsSummariesListNextResponse> {
     return this.client.sendOperationRequest(
-      { scope, nextLink, options },
+      { resourceScope, nextLink, options },
       listNextOperationSpec
     );
   }
@@ -445,7 +449,7 @@ const listByReservationOrderAndReservationOperationSpec: coreClient.OperationSpe
   serializer
 };
 const listOperationSpec: coreClient.OperationSpec = {
-  path: "/{scope}/providers/Microsoft.Consumption/reservationSummaries",
+  path: "/{resourceScope}/providers/Microsoft.Consumption/reservationSummaries",
   httpMethod: "GET",
   responses: {
     200: {
@@ -464,7 +468,7 @@ const listOperationSpec: coreClient.OperationSpec = {
     Parameters.reservationId1,
     Parameters.reservationOrderId1
   ],
-  urlParameters: [Parameters.$host, Parameters.scope],
+  urlParameters: [Parameters.$host, Parameters.resourceScope],
   headerParameters: [Parameters.accept],
   serializer
 };
@@ -518,7 +522,11 @@ const listNextOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.ErrorResponse
     }
   },
-  urlParameters: [Parameters.$host, Parameters.scope, Parameters.nextLink],
+  urlParameters: [
+    Parameters.$host,
+    Parameters.nextLink,
+    Parameters.resourceScope
+  ],
   headerParameters: [Parameters.accept],
   serializer
 };