+++ sdk/consumption/arm-consumption/CHANGELOG.md
@@ -1,15 +1,34 @@
 # Release History
+    
+## 10.0.0 (2023-02-03)
+    
+**Features**
 
-## 9.2.1 (Unreleased)
+  - Added operation LotsOperations.listByCustomer
+  - Added Interface LotsListByCustomerNextOptionalParams
+  - Added Interface LotsListByCustomerOptionalParams
+  - Added Interface ModernReservationRecommendationProperties
+  - Added Interface ModernSharedScopeReservationRecommendationProperties
+  - Added Interface ModernSingleScopeReservationRecommendationProperties
+  - Added Type Alias LotsListByCustomerNextResponse
+  - Added Type Alias LotsListByCustomerResponse
+  - Added Type Alias ModernReservationRecommendationPropertiesUnion
+  - Interface CreditSummary has a new optional parameter eTagPropertiesETag
+  - Interface LegacyChargeSummary has a new optional parameter azureMarketplaceCharges
+  - Interface LegacyUsageDetail has a new optional parameter benefitId
+  - Interface LegacyUsageDetail has a new optional parameter benefitName
+  - Interface ModernChargeSummary has a new optional parameter subscriptionId
+  - Interface ModernReservationRecommendation has a new optional parameter resourceType
+  - Enum KnownEventType has a new value CreditExpired
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Operation ReservationRecommendationDetails.get has a new signature
+  - Interface BudgetFilter no longer has parameter not
+  - Interface LegacyChargeSummary no longer has parameter marketplaceCharges
+  - Parameter scope of interface ModernReservationRecommendation is now required
+    
+    
 ## 9.2.0 (2023-01-03)
     
 **Features**